requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
python-dotenv
openpyxl
requests
gunicorn
